LPS-96279 Conjunction change is now independent from SegmentEdit
